Move repository implementation to data layer
