+++ app/src/test/java/com/victormordur/gihbli/app/data/repository/FilmRepositoryTest.kt
@@ -1,4 +1,4 @@
-package com.victormordur.gihbli.app.domain.repository
+package com.victormordur.gihbli.app.data.repository
 
 import com.victormordur.gihbli.app.domain.model.Film
 import com.victormordur.gihbli.app.data.store.DatastoreContract
